Removed IActor.
IActor was removed and all references to it were replaced with Actor.
Actor is the only implementation of IActor and it is tightly coupled
with FPE so there seems little point in having IActor would could
potentially mislead someone later on into thinking they can implement
make another implementation of IActor without a lot of fiddling with the
physics engine.